Introducing configurator package in vdsm-tool
This package union libvirt and sanlock configuration operations for now.
The package defines ModuleConfigure interface and API for modules configuration
tool. The format for the configure command is modified to:

vdsm-tool configure --module libvirt --force

This module also provides is-configured and validate-config verbs with
the same syntax.

--force flag means the configure flow will stop all related services if
needed. Without force, if one of the related services is running, the
configure will fail.

For reuse and easier additional configurers for other services, all the
developer needs to add is an object that inherit ModuleConfigure and
implements the interface for test, dry and configure, then adding the
class to configurers.

Change-Id: I16bf5894e7e55a84b4c2a0caacde383ae7c19242
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/20100
Reviewed-by: Dan Kenigsberg <danken@redhat.com>