lvm: ignore skipped clustered vgs
In VDSM we use lvm locking_type 1 (local flocks on files in locking_dir)
because the clustered vg access is handled with the SPM-HSM logic.
When locking_type is configured to 1 the clustered vgs are skipped for
all the operations (as they need locking_type 2) and by default the exit
code of the lvm command is 5 triggering errors in our module.
Adding the --ignoreskippedcluster flag to lvm commands we avoid the exit
code error in the event of any clustered vg being present in some of the
attached LUNs.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=820982
Change-Id: Idcabb5c1c19106f0db1385787251d951e12ab0cf
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/20248
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>