vdsm: allow hooks to pass down dictionaries in json format
This will allow to pass, and recover network settings in the
*_network_setup hook as json dictionaries using the same
method that is used for XML. In this case the environment
variable _hook_json is used.

before_network_setup hook exposes the requested network
configuration that can be patched by the hook.
after_network_setup exposes the final system network
configuration.

Bug-Url: https://bugzilla.redhat.com/1058707
Change-Id: Ie07c511e9740fd19a1c27baf87e91f9a427d0dcd
Signed-off-by: Miguel Angel Ajo <miguelangel@ajo.es>
Reviewed-on: http://gerrit.ovirt.org/20330
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Miguel Angel Ajo Pelayo <majopela@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/23660
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>