Replace unused getFileList() with more flexible getFileStats()
1. Rename to getFileStats().

New name getFileStats() better describes the essence of that API.
This is safe to rename it because this API is not in use.

2. Add caseSensitive option.

Currently Engine is using getIsoList() and getFloppyList() APIs to
retrieve the list of iso and floppy files respectively. Both these APIs
are working in case-insensitive manner but don't provide file
statistics. getFileStats() returns file statistics. Adding
caseSensitive option allows using getFileStats() instead of
getIsoList() and getFloppyList().

Change-Id: I550827c7b4c7e11fe09e41745fcc9d91249c6c23
Bug-Url: https://bugzilla.redhat.com/1005889
Signed-off-by: Sergey Gotliv <sgotliv@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/20476
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-by: Ayal Baron <abaron@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>