Deprecate volume mtime
In pre-historic 2.y days, Engine used volume's "mtime" attribute to
deduces the ancestry of a volume. See for example BZ#513396.
In case of an error reading the metadata, "0" or "" was returned.
This was a bad heuristic that was fixed to using proper pointer to
parents and children.

Engine 3.y continues to use the mtime field only when importing ancient
images that lack proper creation time in their OVF.

Maintaining this attribure requires costly readings of volume metadata.

This patch drops the mtime-related code. For backward compatibility with
old Vdsms and Engines that may expect the existance of this attribute,
we keep writing "0" to the metadata and reporting "0" to Engine.

Change-Id: I38b0a636222fa74125d25f0c3c9ea5b3e5701565
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/20504
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>