cleanup: Improve lib/vdsm imports (PEP328)
PEP328 ( http://www.python.org/dev/peps/pep-0328/ ) does away with
the ambiguity of where the modules that are being imported reside
making the codebase much more browsable.

With a codebase of:
    package/
        __init__.py
        subpackage1/
            __init__.py
            moduleX.py
            moduleY.py
        subpackage2/
            __init__.py
            moduleZ.py
        moduleA.py

this (moduleY.py):
    import moduleX
    import moduleZ
becomes
    from . import moduleX
    from ..subpackage2 import moduleZ

Which makes it much more evident where things are and avoids possible
shadowing of packages in site-packages.

Change-Id: I249cfa0ad734ea45ecbbecbade9daeed6c3adc12
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/20555
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>