xmlrpc: Parsing error logging enhancement - vdsClient
Logging enhancement which helps understand the issues during
parsing xml response obtained from vdsm. Parsing
issues can occur when bad characters are in xml etc.

In order to get raw response we need to wrap one of transport
classes (plain or provided class for ssl) and present it to
the user when ExpatError is raised.

Change-Id: I0c39ac88e82f2e511f0025f18586e7a251a17321
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=982065
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/21506
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Sergey Gotliv <sgotliv@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>