xmlrpc: Parsing error logging enhancement - vdsClient
Logging enhancement which helps understand the issues during
parsing xml response obtained from vdsm. Parsing
issues can occur when bad characters are in xml etc.

In order to get raw response we need to wrap one of transport
classes (plain or provided class for ssl) and present it to
the user when ExpatError is raised.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=982065
Change-Id: Ife29c4f7749b9cd8a4ad892f486d91509e505ae4
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: http://gerrit.ovirt.org/20627
Reviewed-by: Dan Kenigsberg <danken@redhat.com>