storage: fix getDomUuidFromVolumePath()
getDomUuidFromVolumePath returned the imgUUID instead of the sdUUID.

Since this is used as the parameter for getProcessPool() a process pool
was created per image id instead of per domain id.

Since imgUUIDs are much more abundant this can cause the creation of an
obscene number of process pools and in turn a lot of processes.

Change-Id: I17267993c8e3a582b1277f1e8aacd6bc69c8e4ec
Bug-Url:http://bugzilla.redhat.com/017735
Signed-off-by: Saggi Mizrahi <smizrahi@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/20747
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Tested-by: Federico Simoncelli <fsimonce@redhat.com>