Quarantine ovirt-3.0.0 network upgrade
Since ovirt-3.0.z, Vdsm uses libvirt networks (with names vdsm-*) to store
its own networks. Older Vdsms did not have those defined, and used only
Linux bridges. _syncLibvirtNetworks created a vdsm-bla network for every
bridge named bla, even if that bridge had nothing to do with ovirt.

This patch replaces _syncLibvirtNetworks with an upgrade that does
this conversion on the rare occasion that upgrade is needed. It
identifies this occasion by the existence of ovirtmgmt bridge
and nonexistence of a vdsm-ovirtmgmt network.

Moving the code to a separate upgrade would make it easier to get rid of
it once ovirt-3.0.0 hosts are no longer supported.

Change-Id: Icb2ffff1715dc3b92fc6c198dbb4c49f0bbef0cb
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/20803