sampling.ImagePathStatus: drop unused code
In pre-historic rhev-2.1 days, ImagePathStatus thread was used to
collect and cache information about the image repository accessibility.
However, this code was never used by ovirt-3.y.

By default, images_check_times is set to 0, meaning that the
sampling.ImagePathStatus thread is never started. It is highly unlikely
that there's a secret vdsm client out there that explicitly turns on
this monitoring thread and read the 'storageDomains' element out of
getVdsStats.

Change-Id: I682d9730ac087e0a24ba5c72dbac912c87a6452f
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/20812
Reviewed-by: Petr Šebek <psebek@redhat.com>
Tested-by: Petr Šebek <psebek@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>