Transmogrify netinfo gateway default value from 0.0.0.0 to ''
Antecedently, if a device lacked IP addressing information,
netinfo proclaimed the ipaddr as '', netmask as ''
and gateway as '0.0.0.0'.

The transpose from '0.0.0.0' to '' is momentous both for congruity
and for input corroboration during setupNetworks. setupNetwrosk would
fail for an (ip: '', netmask: '', gateway: '0.0.0.0'). This
can bechance if the unified network persistence upgrade was
perpetrated on a network with no ip configuration.

The next time I get a -1 on a commit message I promise to
reword it to Biblical English (Present work excluded).

Change-Id: I64ec08bb57f317b30038c3b53851d6a9a4a90e86
Signed-off-by: Assaf Muller <amuller@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/20912
Reviewed-by: Dan Kenigsberg <danken@redhat.com>