fuserTests fix for f20 and arch
In newer distributions, fuser doesn't report a process as accessing
it's own executable file, making this test fail. We don't know why
this has changed (tried with newer kernels and newer fuser and the
issue persists) but we do not care enough about fuser reporting all
the pids from the processes created and alive of an executable file.
We fix this test for the fuser use case that we have.

Change-Id: I9ead0d5dbed6a9373f1437f6484ba2f2ce108c2b
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/20921
Reviewed-by: Dan Kenigsberg <danken@redhat.com>