Improve error when multipath can't access a pv
InvalidPhysDev gave a cryptic error message - Invalid physical device.
The message was amended to "Multipath cannot access physical device(s)" along
with the device(s) name(s). The exception name was changed according to the new
message content. Note: The only thing needed here by the oVirt engine is the
error code, which was not changed in this patch, so there are no
compatibility issues with older oVirt engines.

Change-Id: I648ee519873c51573e6e6306b79380f54bb25d2e
Bug-url: https://bugzilla.redhat.com/852003
Signed-off-by: Vered Volansky <vvolansk@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/21089
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>