Remove onboot, delay, forward_delay from setupNetworks API
The engine never sends onboot and delay so they are safe to remove.
forward_delay is an undocumented API key that was deemed
harmful and unneeded and is removed as well. All three
keys are now blacklisted and will not be written to ifcfg
files if sent via the API. Instead, onboot will always be
yes and delay will always be 0.

Change-Id: Ib3f3d35f13e23947c702f3573b6bb83d444356b2
Signed-off-by: Assaf Muller <amuller@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/20967
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>