gluster prepareImage: return gluster-specific information
Since the introduction of gLuster storage domains (1d90895cdda8,
committed 9 months ago), prepareImage returns an additional element
beyond the path to the leaf: a dictionary keyed by "info", carrying
gluster-specific information.

This dictionary has been dropped by a recent refactoring of prepareImage
since the computation of its contents assumes that the leaf volume is
part of a storage pool, and this is harmful to the "Hosted Engine"
effort.

This patch re-introduces the lost functionality, in the price of a
redundant production of a volume. To make this a little less expensive
and harmful to HE, we limit the production to gluster volumes.

This patch is not the child we've prayed for, but it works.

Special thanks to Samuli Heinonen <samppah@neutraali.net> for reporting
the bug and testing this patch.

Bug-Url: https://bugzilla.redhat.com/1022961
Change-Id: If9a8fbf2baafa1f9bfc3677b1301c7934ca74651
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Signed-off-by: Eduardo Warszawski <ewarszaw@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/21059
Reviewed-by: Ewoud Kohl van Wijngaarden <ewoud@kohlvanwijngaarden.nl>
Reviewed-by: Sandro Bonazzola <sbonazzo@redhat.com>