ipwrapper: Add ovs, libteam and vlan.device reporting support
This patch enables proper classification of ovs bridges and
teamd "bonds" as well as adds one extra attribute to vlans and
macvlans, device, that reports upon which device they are created.

Additionally, it adds memoization to detectType so successive
calls don't need to go all the way down to cpopen for ethtool. This
is okay since it is very rare, unadvisable and probably unsupported
to change device names on the fly and use those vacant names for
other device types.

Change-Id: Ie96296fd59b1ef1a8cea3dcdee6ac6f9c33de0da
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/21078
Reviewed-by: Dan Kenigsberg <danken@redhat.com>