Add iproute2 configurator
This patch adds a new host network configurator based on iproute2.
We could just reuse the ifcfg's configurator totally and add a new
ConfigWriter based on iproute2. But it turns out this approach is
not very suitable to the iprtoute2 model. The ifcfg configurator
depends on the behavior of ifup/ifdown, so it imposes some
unnecessary restriction on the new ConfigWriter.  So this patch
chooses to add a new configurator and reuse some code of ifcfg
configurator. In future, we could move those common code to a base
class after refactoring ifcfg configurator.

This patch dosen't include the support for dhcp and ipv6, which
will be added in following patches.

Change-Id: I762cfa30f78c5a46507b86f53e98bcf79dfc5844
Signed-off-by: Mark Wu <wudxw@linux.vnet.ibm.com>
Reviewed-on: http://gerrit.ovirt.org/15301
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>