domainMonitor: tag pool monitored domains
As part of the monitoring implementation:

 7b1cc6a Adding [start|stop]MonitoringDomain()

updateMonitoringThreads has been modified to assume that the list of
the monitored domains is the same as the list of the attached domains:

+        poolDoms = self.getDomains()
...
-        for sdUUID in monitoredDomains:
+        for sdUUID in poolDoms:
             if sdUUID not in activeDomains:
                 try:
                     self.domainMonitor.stopMonitoring(sdUUID)

This assumption is not valid when we are detaching a storage domain
(as it won't be listed by getDomains anymore). The resulting issue is
that the monitor domain thread is not stopped and the host id is left
acquired (ids file/device kept open by sanlock).

This patch reverts the updateMonitoringThreads changes and adds a tag
mechanism for the pool monitored domains.

Change-Id: I71714b24b7c8bbb437cbb2e8e254690215df020a
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/21165
Reviewed-by: Ayal Baron <abaron@redhat.com>