vm: refresh raw disk before live extension
As part of the online disk extension (raw format on block domains) flow
the leaf logical volume is extended by the SPM to the new size.
On the HSM host (where the VM is running) we need to refresh the volume
in order to update the volume segments and make the new size visible to
the qemu process.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1029501
Change-Id: I4fc3b578a865e25586a719f7c5120df12842d046
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/21177
Reviewed-by: Ayal Baron <abaron@redhat.com>