Allow setting just a single value in MoM policy
This patch prevents a policy coruption that could have happened
when the 01-parameters policy was updated with just a single
value. If there were more values defined, the variables were lost
and the policy ended up broken.

This patch defines the variables in a file that is not touched
by the setMOMPolicyParameters and uses (set name value) syntax
in the generated policy part.

It also remembers the old setting and puts all the values to
the policy part when it is autogenerated.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1025780

Change-Id: I7bb1001a05e670e393b2cc0ab4b271b3b333d2e8
Signed-off-by: Martin Sivak <msivak@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/21225
Reviewed-by: Amador Pahim <apahim@redhat.com>
Reviewed-by: Doron Fediuck <dfediuck@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/21839
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>