sampling: take storageDomains directly from storage subsys
HostStatsThread does not poll storageDomains statistics, but still it is
used to relay them from the storage subsystem to the getVdsStats api.
This is cumbersome and confusing. This patch eliminates the need to pass
a clientIF reference to the HostStatsThread, and keeps the knowledge of
which statistic elements are provided (code, delay, lastCheck, etc)
with the storage subsystem.

Change-Id: I1a86b828fdf43d74bf4475d2c6632c4511519356
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/21334
Reviewed-by: Nir Soffer <nsoffer@redhat.com>