lvm: Allow multiple LVs to be refreshed at once.
This change refactors the refreshLV method to allow multiple
volumes to be refreshed at once as possible with the activate
and deactive methods.

This is patch 1 in a series of 3 that aim to resolve BZ#1009812.

This patch is required by the changes listed below :

lvm: Make the isLVActive() method public.
http://gerrit.ovirt.org/#/c/21386/

blockSD: Ensure active volumes are refreshed before use.
http://gerrit.ovirt.org/#/c/21387

bug-url: https://bugzilla.redhat.com/1009812
Change-Id: I1ecc64b8ca0133b030ba5bfa37f1a2c55067dd5d
Signed-off-by: Lee Yarwood <lyarwood@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/19871
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Tested-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>