Make ip link wrapper hidden config be at class lvl
Before this patch we got the config for each device that needed it.
This allowed each device to pick up changes in the config that
could have happened between each device object creation, at the
cost of speed and simplicity. Since config is supposed to be
stable during the life of a vdsm process, this was moved to the class
level.

Change-Id: I43e484c2e2614b1d9bbd67468e3eb050d55d32f6
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/21407
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>