Peform related services restart after reconfiguring
Although we tried to avoid doing restart to external services as part of
the pre-start of vdsm, the full implementation of vdsm-tool API that allows to
check isconfigured before performing the configuration itself is missing in
ovirt-3.3.

Currently when running libvirt-configure we return 0 (SUCCESS) on both
reconfigure and if libvirt conf is already configured with vdsm details.
That's lead to a problem to decide when to stop the pre-start part and wait
for manual restart, or continue as it is already working with the right
configurations.

To avoid more changes, we return the auto restart to allow easy start of
vdsm without more specific tweaks.

Change-Id: I8216f9556e6ef4df96977db494c5b5169bfae80a
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1031592
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/21356
Reviewed-by: Alon Bar-Lev <alonbl@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>