netconf: Add limit to rollback for recursive configurator
This is the first patch of a series that is intended to provide
a default configurator memory rollback implementation and define
the contracts of the rollback operation. To do that, first it is
necessary to define a difference between a regular setupNetworks
operation and one performed with the purpose of rolling back.

If a configurator implemented the rollback operation via calling
setupNetworks, we would potentially end up recursing until it
succeeded (and we have no guarantee for success).

This patch adds a parameter to the configurator constructor and
the setupNetworks options that allows the configurator to pass
to setupNetworks inRollback=True when it uses it for rollback
purposes.

Change-Id: Ibd24cfb15f37695d999c66deada9c6dcd6bef574
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/21615
Reviewed-by: Dan Kenigsberg <danken@redhat.com>