vm: Fix vm unpausing during recovery
Domain monitor thread was not reporting when going from unknown state to
known state. The thread would start with a bogus valid state, before
checking the domain. If the first domain check found the domain as
valid, no notification was emitted. If there were paused vms while
connecting to the pool, they were not unpaused as they should.

This patch fix various modules:

- Domain monitor status lastCheck is initialized now to special
  NOT_CHECKED_YET value. The monitor thread use this value to detect the
  first sample, and emit the missing notification.

- To get the new event when monitor thread starts, domain state change
  are registered now *before* the monitor threads are started.

- To get the vm unpause code to work in recovery mode, the Vm.sdIds is
  initialized in Vm._run. This probably fixes other issues with
  recovered vms.

Change-Id: I56c468edb967de5bcf15259029ace84333fc4b7e
Bug-Url: https://bugzilla.redhat.com/1003588
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/21649
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>