netconf: provide a default rollback for configurators
This patch adds a default rollback method that configurators are
free to override as long as they keep the contract in the docstring
of the vdsm/netconf/__init__.py:rollback method, i.e., return None
(even implicitly, not returning anything) when there is no need for
rollback at tha API.py level. If there is something to rollback,
return it in a Config object.

The Ifcfg configurator, being on its way to deprecation is supposed
to keep its current memory rollback, so we can keep its overriding
method. The iproute2 configurator won't need to declare it.

Change-Id: I267e7d682e5075695fbc16fd45b241c4137296da
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/21739
Reviewed-by: Dan Kenigsberg <danken@redhat.com>