F20 netinfo: fix linkShowDev parsing in getVlanID
Newer linux distributions return more detailed information for each
link. Where before we had:

    3: enp3s0.11@enp3s0: <BROADCAST,MULTICAST,UP,LOWER_UP> mtu 1500 qdisc noqueue state UP mode DEFAULT group default
        link/ether c8:60:00:99:01:37 brd ff:ff:ff:ff:ff:ff promiscuity 0
        vlan id 11 <REORDER_HDR>

Now we have:

    3: enp3s0.11@enp3s0: <BROADCAST,MULTICAST,UP,LOWER_UP> mtu 1500 qdisc noqueue state UP mode DEFAULT group default
        link/ether c8:60:00:99:01:37 brd ff:ff:ff:ff:ff:ff promiscuity 0
        vlan protocol 802.1Q id 11 <REORDER_HDR>

Thus, since after the vlan type indication we can have several key, value
pairs, it is necessary to change the parsing so that it looks at the keys
to decide if it's the value we want.

This patch is a very very partial backport of the master branch fix:
    http://gerrit.ovirt.org/#/c/21054/
That patch was very complex and introduced different behaviors that
are not suitable for a backport such as the one this patch is doing.

Change-Id: I6fbc492d2eee88e856e893ad69fe7552ede30b1a
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/21759
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>