sp: refresh metadata on hsm when listing domains
Previously to get the most updated domains list we were issuing an
invalidateMetadata request before getDomains (updateMonitoringThreads).

Given that invalidateMetadata is ignored on SPM (as the cache of the
single writer is always up to date) we can encapsulate the metadata
read in getDomainsMap (useful to remove the pool metadata) and always
call invalidateMetadata before reading PMDK_DOMAINS.
This will also ensure that on HSM the domains list is always up to
date.

The @unsecured decorator is removed from validatePoolSD and
validateAttachedDomain to make explicit that they should be used only
on the SPM (as HSM should never use these validations), and therefore
getDomains won't trigger a metadata invalidation from there.

Change-Id: I095cd0760076fb4be97a776498af78a40ff84112
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/21786
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>