sp: refactor out the metadata access from StoragePool
All the storage pool metadata operations have been relocated to a
specific backend: StoragePoolDiskBackend. It is now possible,
implementing the required API, to use different backends for the
pool metadata.

A backend may require other attributes and methods in addition to
the regular API used by StoragePool, these specific features should
be handled outside of the StoragePool implementation, typically in
the HSM class where the backward compatibility layer is maintained.

Change-Id: I75493d1db60e51cccd5231b516f963c970d24c99
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/22132
Reviewed-by: Dan Kenigsberg <danken@redhat.com>