lvm.env: move to /usr/share/vdsm/storage
lvm.env is a tiny helper script, intended to make it a bit easier to
debug Vdsm-controlled lvm state. It is being created whenever the lvm
module is loaded, which generates an annoying warning message in unit
test.

This patch makes lvm.env ship statically with vdsm. It can be used by

    . /usr/share/vdsm/storage/lvm.env

Change-Id: I6ae6bc789efa22321eb06422e910bf272c575007
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/21872
Reviewed-by: Nir Soffer <nsoffer@redhat.com>