vm: Refix vm unpausing during recovery
Commit 45f3037ca1 should have fixed vm unpausing during recovery, but it
did not because of unfortunate cherry-pick.  The original patch was
initializing vm sdIds list when running a vm in recovery mode. This code
path was removed in commit 93d09f6567 which was merged on the same day.

This patch adds back the missing code path and remove leftover comment
about recovery mode.

Change-Id: Ibbfbed22f1787f727884eeed72d2b15835ebe2dc
Relates-To: http://gerrit.ovirt.org/21649/
Bug-Url: https://bugzilla.redhat.com/1036358
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/21892
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Sergey Gotliv <sgotliv@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>